Refactor SubAccountClient to use pydantic types
