+++ src/pypaystack2/paystack.py
@@ -18,7 +18,7 @@ from pypaystack2.sub_clients.async_clients.products import AsyncProductClient
 from pypaystack2.sub_clients.async_clients.refunds import AsyncRefundClient
 from pypaystack2.sub_clients.async_clients.settlements import AsyncSettlementClient
 from pypaystack2.sub_clients.async_clients.splits import AsyncTransactionSplitClient
-from pypaystack2.sub_clients.subaccounts import AsyncSubAccountClient, SubAccountClient
+from pypaystack2.sub_clients.async_clients.subaccounts import AsyncSubAccountClient
 from pypaystack2.sub_clients.subscriptions import (
     AsyncSubscriptionClient,
     SubscriptionClient,
@@ -61,6 +61,9 @@ from pypaystack2.sub_clients.sync_clients.settlements import (
 )
 from pypaystack2.sub_clients.sync_clients.splits import (
     TransactionSplitClient,
+)
+from pypaystack2.sub_clients.sync_clients.subaccounts import (
+    SubAccountClient,
 )
 from pypaystack2.sub_clients.terminals import TerminalClient, AsyncTerminalClient
 from pypaystack2.sub_clients.transactions import (
