+++ tests/test_paystack.py
@@ -45,7 +45,7 @@ from pypaystack2.sub_clients.async_clients.plans import AsyncPlanClient
 from pypaystack2.sub_clients.async_clients.refunds import AsyncRefundClient
 from pypaystack2.sub_clients.async_clients.settlements import AsyncSettlementClient
 from pypaystack2.sub_clients.async_clients.splits import AsyncTransactionSplitClient
-from pypaystack2.sub_clients.subaccounts import AsyncSubAccountClient
+from pypaystack2.sub_clients.async_clients.subaccounts import AsyncSubAccountClient
 from pypaystack2.sub_clients.subscriptions import AsyncSubscriptionClient
 from pypaystack2.sub_clients.sync_clients.products import AsyncProductClient
 from pypaystack2.sub_clients.terminals import AsyncTerminalClient
